updated footer link to advanced search copy
