+++ src/_scss/pages/covid19/_linkToAdvancedSearch.scss
@@ -1,0 +1,24 @@
+@import "components/_footerLinkToAdvancedSearch";
+
+.footerLinkToAdvancedSearch__content_covid {
+    // Match padding in the Data Sources & Methodology section
+    @include media($medium-screen) {
+        padding-right: 8%;
+        padding-left: 8%;
+    }
+    button {
+        @include button-link;
+        font-size: 1.8rem;
+        color: $color-white;
+
+        &:hover {
+            text-decoration: underline;
+            color: $color-white;
+        }
+    }
+
+    a {
+        color: $color-white;
+        font-weight: $font-normal;
+    }
+}
