+++ src/js/containers/covid19/Covid19Container.jsx
@@ -18,7 +18,7 @@ import { LoadingWrapper } from 'components/sharedComponents/Loading';
 // import { Picker } from 'data-transparency-ui';
 import ShareIcon from 'components/sharedComponents/stickyHeader/ShareIcon';
 // import { defaultSortFy } from 'components/sharedComponents/pickers/FYPicker';
-import FooterLinkToAdvancedSearchContainer from 'containers/shared/FooterLinkToAdvancedSearchContainer';
+import LinkToAdvancedSearchContainer from 'containers/covid19/LinkToAdvancedSearchContainer';
 import RedirectModalContainer from 'containers/redirectModal/RedirectModalContainer';
 import { covidPageMetaTags } from 'helpers/metaTagHelper';
 import BaseOverview from 'models/v2/covid19/BaseOverview';
@@ -26,12 +26,8 @@ import { jumpToSection, latestSubmissionDateFormatted } from 'helpers/covid19Hel
 import {
     slug,
     getEmailSocialShareData,
-    scrollPositionOfSiteHeader,
-    footerTitle,
-    footerDescription
+    scrollPositionOfSiteHeader
 } from 'dataMapping/covid19/covid19';
-import { initialState as defaultAdvancedSearchFilters, CheckboxTreeSelections } from 'redux/reducers/search/searchFiltersReducer';
-import { applyStagedFilters } from 'redux/actions/search/appliedFilterActions';
 import { fetchDEFCodes, fetchOverview, fetchAllSubmissionDates } from 'helpers/disasterHelper';
 import { setDEFCodes, setOverview, setLatestSubmissionDate } from 'redux/actions/covid19/covid19Actions';
 import { showModal } from 'redux/actions/redirectModal/redirectModalActions';
@@ -96,19 +92,6 @@ const Covid19Container = () => {
             }
         };
     }, [defCodes]);
-
-    const addDefCodesToAdvancedSearchFilter = () => dispatch(applyStagedFilters(
-        Object.assign(
-            {}, defaultAdvancedSearchFilters,
-            {
-                defCodes: new CheckboxTreeSelections({
-                    require: defCodes.map((code) => code.code),
-                    exclude: [],
-                    counts: [{ value: "COVID-19", count: defCodes.length, label: "COVID-19 Response" }]
-                })
-            }
-        )
-    ));
 
     useEffect(() => {
         const getAllSubmissionDates = async () => {
@@ -207,10 +190,7 @@ const Covid19Container = () => {
                         <section className="body__section" id="covid19-data_sources_and_methodology">
                             <DataSourcesAndMethodology
                                 handleExternalLinkClick={handleExternalLinkClick} />
-                            <FooterLinkToAdvancedSearchContainer
-                                title={footerTitle}
-                                description={footerDescription}
-                                onClick={addDefCodesToAdvancedSearchFilter} />
+                            <LinkToAdvancedSearchContainer />
                         </section>
                     </div>
                     <RedirectModalContainer />
