+++ src/js/containers/covid19/LinkToAdvancedSearchContainer.jsx
@@ -1,0 +1,54 @@
+/**
+ * LinkToAdvancedSearchContainer.jsx
+ * Created by Lizzie Salita 7/28/20
+ */
+
+import React from 'react';
+import { useDispatch, useSelector } from 'react-redux';
+import Router from 'containers/router/Router';
+import { isEqual } from 'lodash';
+import { clearAllFilters } from 'redux/actions/search/searchFilterActions';
+import { resetAppliedFilters, applyStagedFilters } from 'redux/actions/search/appliedFilterActions';
+import { initialState as defaultAdvancedSearchFilters, CheckboxTreeSelections } from 'redux/reducers/search/searchFiltersReducer';
+
+const FooterLinkToAdvancedSearchContainer = () => {
+    const dispatch = useDispatch();
+    const defCodes = useSelector((state) => state.covid19.defCodes, isEqual);
+
+    console.log(defCodes);
+    const addDefCodesToAdvancedSearchFilter = () => dispatch(applyStagedFilters(
+        Object.assign(
+            {}, defaultAdvancedSearchFilters,
+            {
+                defCodes: new CheckboxTreeSelections({
+                    require: defCodes.map((code) => code.code),
+                    exclude: [],
+                    counts: [{ value: "COVID-19", count: defCodes.length, label: "COVID-19 Response" }]
+                })
+            }
+        )
+    ));
+
+    const clickedSearch = () => {
+        dispatch(clearAllFilters());
+        dispatch(resetAppliedFilters());
+        addDefCodesToAdvancedSearchFilter();
+        Router.history.push('/search');
+    };
+
+    return (
+        <div className="footerLinkToAdvancedSearch">
+            <div className="footerLinkToAdvancedSearch__content footerLinkToAdvancedSearch__content_covid">
+                <h4>Looking for more information about COVID-19 federal awards?</h4>
+                <p>
+                    Visit our <button onClick={clickedSearch}>Advanced Search</button> page to see individual awards funded through the COVID-19 response and keep an eye out for purple COVID-19 badges found throughout the site.
+                </p>
+                <p>
+                    Visit our <a href="#/download_center/custom_account_data">Custom Account Data</a> download for COVID-19 award spending data with a higher degree of granularity.
+                </p>
+            </div>
+        </div>
+    );
+};
+
+export default FooterLinkToAdvancedSearchContainer;
