+++ src/js/dataMapping/covid19/covid19.js
@@ -17,9 +17,6 @@ export const scrollPositionOfSiteHeader = (cookie) => (cookie ? 96 : 187);
 export const defCodes = ['L', 'M', 'N', 'O', 'P'];
 export const allDefCAwardTypeCodes = ['02', '03', '04', '05', '06', '07', '08', '09', '10', '11'];
 
-export const footerTitle = 'Looking for more insight about COVID-19 Spending?';
-export const footerDescription = 'for more in-depth analysis on this COVID-19 and more';
-
 export const amountsHeight = 400;
 export const sankeyHeight = 400;
 export const amountsPadding = {
